backend/app/main.py
<CODE_BLOCK>
```python
import os
from fastapi import FastAPI, Depends, HTTPException, status
from fastapi.security import HTTPBearer
from typing import Dict, List, Any
from jose import jwt, JWTError
import requests

app = FastAPI()

# --- Keycloak Authentication Configuration ---
# Get Keycloak settings from environment variables
KEYCLOAK_REALM = os.getenv("KEYCLOAK_REALM", "master")
KEYCLOAK_URL = os.getenv("KEYCLOAK_URL", "http://keycloak:8080") # Use service name for Docker internal communication
KEYCLOAK_CLIENT_ID = os.getenv("KEYCLOAK_CLIENT_ID", "account") # Replace with your client ID for the backend gateway
KEYCLOAK_CERTS_URL = f"{KEYCLOAK_URL}/realms/{KEYCLOAK_REALM}/protocol/openid-connect/certs"
KEYCLOAK_ISSUER = f"{KEYCLOAK_URL}/realms/{KEYCLOAK_REALM}"

bearer_scheme = HTTPBearer()
def get_keycloak_public_key():
    """Fetches Keycloak realm public key."""
    try:
        response = requests.get(KEYCLOAK_CERTS_URL)
        response.raise_for_status()
        keys = response.json()['keys']
        # Find the key with the correct algorithm (usually RS256)
        for key in keys:
            if key['use'] == 'sig' and key['kty'] == 'RSA':
                from jose.jwk import construct
                return construct(key)
        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail="Public key not found in Keycloak certificates")
    except requests.exceptions.RequestException as e:
        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail=f"Could not fetch public key from Keycloak: {e}")

def verify_token(token: str = Depends(bearer_scheme)) -> Dict[str, Any]:
    """Verifies Keycloak JWT token and returns the user payload."""
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Could not validate credentials",
        headers={"WWW-Authenticate": "Bearer"},
    )
    try:
        public_key = get_keycloak_public_key()
        payload = jwt.decode(token, public_key, algorithms=["RS256"], audience=KEYCLOAK_CLIENT_ID, issuer=KEYCLOAK_ISSUER)
        return payload
    except JWTError:
        raise credentials_exception
    except Exception as e:
        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail=f"Error decoding token: {e}")

# --- Placeholder Clients for Other Services ---
# Replace with actual client implementations for each service
class OpenMetadataClient:
    def list_datasets(self, user: Dict[str, Any]) -> List[Dict[str, Any]]:
        print(f"OpenMetadataClient: Listing datasets for user {user['username']}")
        # Simulate fetching data
        return [
            {"id": "dataset1", "name": "sales_data", "description": "Sales data"},
            {"id": "dataset2", "name": "customer_churn", "description": "Customer churn data"},
        ]

class PrefectClient:
    def get_pipelines_for_dataset(self, dataset_id: str, user: Dict[str, Any]) -> List[Dict[str, Any]]:
        print(f"PrefectClient: Getting pipelines for dataset {dataset_id} for user {user['username']}")
        # Simulate fetching data
        if dataset_id == "dataset1":
            return [
                {"id": "pipeline1", "name": "daily_sales_sync", "status": "running"},
                {"id": "pipeline2", "name": "weekly_sales_report", "status": "completed"},
            ]
        return []

openmetadata_client = OpenMetadataClient()
prefect_client = PrefectClient()

# --- Root Endpoint ---
@app.get("/")
async def read_root():
    return {"message": "Unified Data Orchestration Platform Backend Gateway"}

# --- Placeholder Endpoints for Other Services ---

# Example endpoint demonstrating integration with OpenMetadata and Prefect
@app.get("/api/datasets", dependencies=[Depends(verify_token)])
async def get_datasets(user: Dict[str, Any] = Depends(verify_token)):
    """
    Fetches datasets from OpenMetadata and enriches them with related pipelines from Prefect.
    """
    datasets = openmetadata_client.list_datasets(user=user)
    for ds in datasets:
        ds["pipelines"] = prefect_client.get_pipelines_for_dataset(ds["id"], user=user)
    return datasets

@app.get("/api/pipelines", dependencies=[Depends(verify_token)])
async def get_pipelines(user: Dict[str, Any] = Depends(verify_token)):
    """
    Placeholder endpoint to get a list of pipelines.
    Integrate with Prefect API here.
    """
    return {"message": "Placeholder for fetching pipelines"}

@app.get("/api/models", dependencies=[Depends(verify_token)])
async def get_models(user: Dict[str, Any] = Depends(verify_token)):
    """
    Placeholder endpoint to get a list of ML models.
    Integrate with MLflow API here.
    """
    return {"message": "Placeholder for fetching ML models"}

@app.post("/api/trigger-airbyte-sync/{connection_id}", dependencies=[Depends(verify_token)])
async def trigger_airbyte_sync(connection_id: str, user: Dict[str, Any] = Depends(verify_token)):
    """
    Placeholder endpoint to trigger an Airbyte sync job.
    Integrate with Airbyte API here.
    """
    return {"message": f"Placeholder for triggering Airbyte sync for connection: {connection_id}"}

@app.get("/api/protected")
async def protected_endpoint(user: Dict[str, Any] = Depends(get_current_user)):
    """
    A protected endpoint that requires authentication.
    """
    return {"message": f"Hello {user.get('preferred_username', 'Authenticated User')}, you are authenticated!"}

# You would implement role-based access control here using the 'user' object and Keycloak roles
# Example: @app.get("/api/admin-only", dependencies=[Depends(has_role("admin"))])

```
</CODE_BLOCK>