backend/requirements.txt
txt
fastapi
uvicorn
requests
python-jose[cryptography]